endpoints de 2. Gestión de Restricciones de Empresa completado

+++ src/main/java/com/example/hackaton_1_mejorada/Domain/restricciones/RestriccionesService.java
@@ -38,6 +38,8 @@ public class RestriccionesService {
     }
 
     public void delete(Long id) {
-        restriccionesRepository.deleteById(id);
+        if (restriccionesRepository.findById(id).isPresent()){
+            restriccionesRepository.deleteById(id);
+        }
     }
 }
